Restructure search/ directory to match cart/ pattern
- Move components to search/components/
- Move services to search/services/
- Move types to search/types/
- Move utils to search/utils/
- Move workers to search/workers/
- Update all import paths to new structure
- Fix external imports in navigation and reports
