+++ ui/src/navigation/components/HeaderContainer.svelte
@@ -2,7 +2,7 @@
   import { getContext, onMount } from "svelte";
   import { cartTotal, uniqueItemCount, isCheckoutSession } from "../../cart/services/CartBusinessService";
   import { ShoppingCart, Menu } from "lucide-svelte";
-  import SearchBar from "../../search/SearchBar.svelte";
+  import SearchBar from "../../search/components/SearchBar.svelte";
   import SidebarMenu from "./SidebarMenu.svelte";
 
   // Import from UI-only store
