+++ ui/src/reports/components/CategoryReportsAdmin.svelte
@@ -11,7 +11,7 @@
     } from "lucide-svelte";
     import { clickable } from "../../shared/actions/clickable";
     import { uploadService } from "../../services/DHTUploadService";
-    import { IndexGenerationService } from "../../search/IndexGenerationService";
+    import { IndexGenerationService } from "../../search/services/IndexGenerationService";
 
     // No longer need DataManager context - using direct imports
