+++ ui/src/search/services/SearchCacheService.ts
