+++ ui/src/search/services/SearchInitializer.ts
