+++ ui/src/search/types/search-types.ts
