+++ ui/src/search/utils/query-utils.ts
