+++ ui/src/search/utils/search-constants.ts
