+++ ui/src/search/workers/embedding-worker.ts
